Move trim to folder
