+++ crates/nu-cli/src/commands/str_/mod.rs
@@ -12,9 +12,6 @@ mod to_datetime;
 mod to_decimal;
 mod to_integer;
 mod trim;
-mod trim_base;
-mod trim_left;
-mod trim_right;
 mod upcase;
 
 pub use capitalize::SubCommand as StrCapitalize;
@@ -30,7 +27,7 @@ pub use substring::SubCommand as StrSubstring;
 pub use to_datetime::SubCommand as StrToDatetime;
 pub use to_decimal::SubCommand as StrToDecimal;
 pub use to_integer::SubCommand as StrToInteger;
-pub use trim::SubCommand as StrTrim;
-pub use trim_left::SubCommand as StrTrimLeft;
-pub use trim_right::SubCommand as StrTrimRight;
+pub use trim::Trim as StrTrim;
+pub use trim::TrimLeft as StrTrimLeft;
+pub use trim::TrimRight as StrTrimRight;
 pub use upcase::SubCommand as StrUpcase;
