+++ crates/nu-cli/src/commands/str_/trim/trim.rs
@@ -1,9 +1,9 @@
-use crate::commands::str_::trim_base::operate;
+use super::operate;
 use crate::commands::WholeStreamCommand;
 use crate::prelude::*;
 use nu_errors::ShellError;
 use nu_protocol::{
-    Signature, SyntaxShape, Value,
+     Signature, SyntaxShape, Value,
 };
 
 pub struct SubCommand;
@@ -11,14 +11,14 @@ pub struct SubCommand;
 #[async_trait]
 impl WholeStreamCommand for SubCommand {
     fn name(&self) -> &str {
-        "str ltrim"
+        "str trim"
     }
 
     fn signature(&self) -> Signature {
-        Signature::build("str ltrim")
+        Signature::build("str trim")
             .rest(
                 SyntaxShape::ColumnPath,
-                "optionally trim text from left side by column paths",
+                "optionally trim text by column paths",
             )
             .named(
                 "char",
@@ -29,7 +29,7 @@ impl WholeStreamCommand for SubCommand {
     }
 
     fn usage(&self) -> &str {
-        "trims text from left side"
+        "trims text"
     }
 
     async fn run(
@@ -37,36 +37,35 @@ impl WholeStreamCommand for SubCommand {
         args: CommandArgs,
         registry: &CommandRegistry,
     ) -> Result<OutputStream, ShellError> {
-        operate(args, registry, &trim_left).await
+        operate(args, registry, &trim).await
     }
 
     fn examples(&self) -> Vec<Example> {
         vec![
             Example {
                 description: "Trim whitespace",
-                example: "echo '  Nu shell ' | str ltrim",
-                result: Some(vec![Value::from("Nu shell ")]),
+                example: "echo 'Nu shell ' | str trim",
+                result: Some(vec![Value::from("Nu shell")]),
             },
             Example {
                 description: "Trim a specific character",
-                example: "echo '=== Nu shell ===' | str ltrim -c '=' | str trim",
-                result: Some(vec![Value::from("Nu shell ===")]),
+                example: "echo '=== Nu shell ===' | str trim -c '=' | str trim",
+                result: Some(vec![Value::from("Nu shell")]),
             },
         ]
     }
 }
-
-fn trim_left(s: &String, char_: Option<char>) -> String {
+fn trim(s: &String, char_: Option<char>) -> String {
     match char_ {
-        None => String::from(s.trim_start()),
-        Some(ch) => String::from(s.trim_start_matches(ch))
+        None => String::from(s.trim()),
+        Some(ch) => String::from(s.trim_matches(ch))
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::commands::str_::trim_base::action;
-    use super::{trim_left, SubCommand};
+    use crate::commands::str_::trim::action;
+    use super::{trim, SubCommand};
     use nu_plugin::test_helpers::value::string;
     use nu_source::Tag;
 
@@ -78,19 +77,20 @@ mod tests {
     }
 
     #[test]
-    fn trims_whitespace_from_left() {
-        let word = string(" andres ");
-        let expected = string("andres ");
+    fn trims() {
+        let word = string("andres ");
+        let expected = string("andres");
 
-        let actual = action(&word, Tag::unknown(), None, &trim_left).unwrap();
+        let actual = action(&word, Tag::unknown(), None, &trim).unwrap();
         assert_eq!(actual, expected);
     }
-    #[test]
-    fn trims_custom_chars_from_left() {
-        let word = string("!!! andres !!!");
-        let expected = string(" andres !!!");
 
-        let actual = action(&word, Tag::unknown(), Some('!'), &trim_left).unwrap();
+    #[test]
+    fn trims_custom_character_both_ends() {
+        let word = string("!#andres#!");
+        let expected = string("#andres#");
+
+        let actual = action(&word, Tag::unknown(), Some('!'), &trim).unwrap();
         assert_eq!(actual, expected);
     }
 }
